Reorganizar tests de categorías según TESTING_STRATEGY.md
- Mover CategoryServiceSpec a test/unit/configdata/CategoryServiceTest
- Mover CategoriesControllerSpec a test/integration/controllers/CategoriesControllerTest
- Extraer StubCategoryService y fixtures a test/fixtures/CategoryFixtures
- Cambiar should→must en bloques de descripción, usar AnyWordSpec+Matchers
